% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_datatable.R
\name{add_width_limit_values_JS}
\alias{add_width_limit_values_JS}
\title{Add value-related Javascript to datatable options.}
\usage{
add_width_limit_values_JS(options, data)
}
\arguments{
\item{options}{A list of datatable options.}

\item{data}{A data.frame.}
}
\value{
The updated list of datatable options.
}
\description{
This function takes a list of datatable options and adds the value-related Javascript from 'value_callback' to these options.
}
