% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_datatable.R
\name{add_with_limit_header_JS}
\alias{add_with_limit_header_JS}
\title{Add header-related Javascript to datatable options.}
\usage{
add_with_limit_header_JS(options, data)
}
\arguments{
\item{options}{A list of datatable options.}

\item{data}{A data.frame.}
}
\value{
The updated list of datatable options.
}
\description{
This function takes a list of datatable options and adds the header-related Javascript from 'header_callback' to these options.
}
