% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_datatable.R
\name{advanced_options}
\alias{advanced_options}
\title{Get advanced datatable options.}
\usage{
advanced_options()
}
\value{
A list of datatable options.
}
\description{
This function returns a list of advanced datatable options.
}
