% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_datatable.R
\name{basic_options}
\alias{basic_options}
\title{Get basic datatable options.}
\usage{
basic_options()
}
\value{
A list of datatable options.
}
\description{
This function returns a list of basic datatable options.
}
