% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_plotly.R
\name{clean_pltly_legend}
\alias{clean_pltly_legend}
\title{Clean the legend of a plotly object}
\usage{
clean_pltly_legend(pltly_obj, new_legend = c())
}
\arguments{
\item{pltly_obj}{A plotly object with a legend to be cleaned.}

\item{new_legend}{An optional vector of strings specifying new legend entries. Default is an empty vector.}
}
\value{
The input plotly object with its legend cleaned.
}
\description{
This function cleans the legend of a plotly object by removing unnecessary duplication.
It is specifically designed to work around a bug that causes facet_wrap to create a separate legend entry for each facet.
}
