% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_shinydashboardplus_adaptions.R
\name{dropdownTabMenu}
\alias{dropdownTabMenu}
\title{dropdownTabMenu function}
\usage{
dropdownTabMenu(
  ...,
  type = c("messages", "notifications", "tasks"),
  title = NULL,
  icon = NULL,
  .list = NULL,
  header = NULL
)
}
\arguments{
\item{...}{additional arguments.}

\item{type}{A character vector of either "messages", "notifications", "tasks". Default is c("messages", "notifications", "tasks").}

\item{title}{The title of the dropdown.}

\item{icon}{The icon to use in the dropdown. If NULL, defaults will be set based on type.}

\item{.list}{A list of items to add to the dropdown.}

\item{header}{The header for the dropdown.}
}
\value{
A dropdown menu in the form of an HTML list.
}
\description{
Dropdown that is actually more of a menu with adapted tasks.
}
\examples{
dropdownTabMenu(type = "messages", title = "Category tab items")
}
