% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_ggplot.R
\name{gantt_plot}
\alias{gantt_plot}
\title{Create a Gantt plot using ggplot and plotly}
\usage{
gantt_plot(df, x, xend, split_var, title, position_label_y)
}
\arguments{
\item{df}{A data frame containing the data to be plotted.}

\item{x}{A string specifying the column name to be used as the x-axis variable.}

\item{xend}{A string specifying the column name to be used as the end of the x-axis variable.}

\item{split_var}{A string specifying the column name to be used as the splitting variable.}

\item{title}{A string specifying the title of the plot.}

\item{position_label_y}{A string specifying the position of y-axis labels.}
}
\value{
A Gantt plot.
}
\description{
This function creates a Gantt plot with the help of ggplot and plotly.
}
