% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_ggplot.R
\name{ggplot_basic_settings}
\alias{ggplot_basic_settings}
\title{Set ggplot basic settings}
\usage{
ggplot_basic_settings()
}
\value{
A list with ggplot settings
}
\description{
Basic ggplot settings are put in a list and returned
}
