% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_ggplot.R
\name{grid_histograms}
\alias{grid_histograms}
\title{Grid histogram}
\usage{
grid_histograms(
  df,
  x,
  color,
  y,
  id,
  y_left = NULL,
  y_right = NULL,
  facet_var = rlang::sym("VIS_Groep"),
  facet_name_var = rlang::sym("VIS_Groep_naam")
)
}
\arguments{
\item{df}{The data frame used to create the plot.}

\item{x}{The variable used on the x-axis of the plot.}

\item{color}{The variable used to color the points or bars in the plot.}

\item{y}{The variable used on the y-axis of the plot.}

\item{id}{The identifier for selecting the data frame source.}

\item{y_left}{The variable used on the left y-axis when creating a comparative plot.}

\item{y_right}{The variable used on the right y-axis when creating a comparative plot.}

\item{facet_var}{The variable used for facet wrapping.}

\item{facet_name_var}{The name of the variable used for facet wrapping.}
}
\value{
A list of ggplot objects.
}
\description{
Function for creating grid histograms.
}
