% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_datatable.R
\name{header_callback}
\alias{header_callback}
\title{Generate Javascript for datatable headers.}
\usage{
header_callback(data = data)
}
\arguments{
\item{data}{An optional data.frame. Default is 'data' in the global environment.}
}
\value{
A character vector containing the JavaScript code.
}
\description{
This function generates Javascript code for datatable headers. The script adjusts header tooltips and names.
The abbreviations have a minimum length of 7 characters.
}
