% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_filters.R
\name{keep_only_relevant_values}
\alias{keep_only_relevant_values}
\title{Keep only relevant values}
\usage{
keep_only_relevant_values(lFilters, sVariable, dfFilters)
}
\arguments{
\item{lFilters}{List of filters to be applied on the data.}

\item{sVariable}{The variable for which relevant values are to be retrieved.}

\item{dfFilters}{Dataframe with the possible filters and values for this dataset}
}
\value{
A list of relevant values for the specified variable.
}
\description{
Filters out only relevant values based on the provided filters
}
\details{
This function removes null elements from the filter list, transforms filter list into elements
suitable for filtering, and retrieves relevant values from the data.
}
