% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_filters.R
\name{limit_n_values_gantt}
\alias{limit_n_values_gantt}
\title{Limit number of values for Gantt chart}
\usage{
limit_n_values_gantt(df, split_var, n_values = 12)
}
\arguments{
\item{df}{A dataframe to be processed}

\item{split_var}{A character vector specifying the variable to be split}

\item{n_values}{An integer specifying the maximum number of values (default is 12)}
}
\value{
A dataframe with a limited number of values for the Gantt chart
}
\description{
This function limits the number of values displayed in a Gantt chart. If the number of distinct
values in the specified variable is less than the limit, the function returns the original dataframe.
}
