% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_datatable.R
\name{make_advanced_table}
\alias{make_advanced_table}
\title{Create an advanced datatable.}
\usage{
make_advanced_table(
  data,
  rownames = FALSE,
  filter = "top",
  extensions = c("Buttons", "ColReorder", "RowReorder"),
  options = advanced_options(),
  limit_width = "values",
  ...
)
}
\arguments{
\item{data}{The data.frame to be displayed.}

\item{rownames}{A logical value indicating whether to display row names.}

\item{filter}{A character string indicating where to display the table filter.}

\item{extensions}{A character vector specifying the DataTables extensions to be used.}

\item{options}{A list of DataTables options.}

\item{limit_width}{A logical value indicating whether to limit column width.}

\item{...}{Additional arguments passed to DT::datatable().}
}
\value{
A datatable object.
}
\description{
This function creates an advanced datatable with various configurable options and features.
}
