% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_datatable.R
\name{make_advanced_table_html}
\alias{make_advanced_table_html}
\title{Create an advanced datatable for HTML rendering.}
\usage{
make_advanced_table_html(data, ...)
}
\arguments{
\item{data}{The data.frame to be displayed.}

\item{...}{Additional arguments passed to 'make_advanced_table()'.}
}
\value{
A datatable object.
}
\description{
This function creates an advanced datatable with options optimized for HTML rendering.
}
