% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_datatable.R
\name{make_basic_table}
\alias{make_basic_table}
\title{Create a basic datatable.}
\usage{
make_basic_table(
  data,
  rownames = FALSE,
  extensions = c("Buttons"),
  options = basic_options(),
  limit_width = "values",
  ...
)
}
\arguments{
\item{data}{The data.frame to be displayed.}

\item{rownames}{A logical value indicating whether to display row names.}

\item{extensions}{A character vector specifying the DataTables extensions to be used.}

\item{options}{A list of DataTables options.}

\item{limit_width}{A character string indicating how to limit column width.}

\item{...}{Additional arguments passed to DT::datatable().}
}
\value{
A datatable object.
}
\description{
This function creates a basic datatable with various configurable options and features.
}
