% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_datatable.R
\name{make_basic_table_html}
\alias{make_basic_table_html}
\title{Create a basic datatable for HTML rendering.}
\usage{
make_basic_table_html(data, ...)
}
\arguments{
\item{data}{The data.frame to be displayed.}

\item{...}{Additional arguments passed to 'make_basic_table()'.}
}
\value{
A datatable object.
}
\description{
This function creates a basic datatable with options optimized for HTML rendering.
}
