% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_gantt.R
\name{module_gantt_ind_server}
\alias{module_gantt_ind_server}
\title{Server function for gantt chart module}
\usage{
module_gantt_ind_server(id, df, df_config_gantt)
}
\arguments{
\item{id}{Module id}

\item{df}{Data frame}

\item{df_config_gantt}{Config data frame for gantt chart}
}
\value{
Shiny module server function
}
\description{
Server function for gantt chart module
}
