% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_gantt.R
\name{module_gantt_ind_ui}
\alias{module_gantt_ind_ui}
\title{UI function for gantt chart module}
\usage{
module_gantt_ind_ui(id, df_config_gantt)
}
\arguments{
\item{id}{Module id}

\item{df_config_gantt}{Config data frame for gantt chart}
}
\value{
Shiny fluidPage UI
}
\description{
UI function for gantt chart module
}
