% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_pickers_special.R
\name{pickerGanttValues}
\alias{pickerGanttValues}
\title{pickerGanttValues function}
\usage{
pickerGanttValues(id, filter_var, df_doorstroom_gantt)
}
\arguments{
\item{id}{A string representing the id of the input element.}

\item{filter_var}{A string representing the variable to filter.}

\item{df_doorstroom_gantt}{A data frame containing the Gantt chart data.}
}
\value{
A pickerInput object.
}
\description{
Function to filter the values in the Gantt chart.
}
