% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_pickers_special.R
\name{pickerGanttVar}
\alias{pickerGanttVar}
\title{pickerGanttVar function}
\usage{
pickerGanttVar(id, element, df_config_gantt, input_var_value = NULL)
}
\arguments{
\item{id}{A string representing the id of the input element.}

\item{element}{A string representing the element.}

\item{df_config_gantt}{A data frame containing the Gantt configuration.}

\item{input_var_value}{A variable value from the input. Default is NULL.}
}
\description{
Function to pick a variable to show values in a Gantt chart.
}
