% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_pickers_special.R
\name{pickerSankeyValues}
\alias{pickerSankeyValues}
\title{pickerSankeyValues function}
\usage{
pickerSankeyValues(id, filter_var, df_sankey, side)
}
\arguments{
\item{id}{A string representing the id of the input element.}

\item{filter_var}{A string representing the variable to filter.}

\item{df_sankey}{A data frame containing the Sankey diagram data.}

\item{side}{A string representing the side of the Sankey diagram.}
}
\value{
A pickerInput object.
}
\description{
Function to pick values for transition of the two Sankey states.
}
