% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_pickers_special.R
\name{pickerSankeyVar}
\alias{pickerSankeyVar}
\title{pickerSankeyVar function}
\usage{
pickerSankeyVar(id, df_sankey, df_config_sankey, state = "left_var")
}
\arguments{
\item{id}{A string representing the id of the input element.}

\item{df_sankey}{A data frame containing the Sankey diagram data.}

\item{df_config_sankey}{A data frame containing the Sankey configuration.}

\item{state}{A string representing the state of the variable. Default is "left_var".}
}
\value{
A pickerInput object.
}
\description{
Function to pick variables for state in a Sankey diagram.
}
