% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_pickers_general.R
\name{pickerSplitVar}
\alias{pickerSplitVar}
\title{pickerSplitVar function}
\usage{
pickerSplitVar(
  id,
  variable = "INS_Splits_variabele",
  name = "color",
  label = "Kleur",
  df
)
}
\arguments{
\item{id}{A string representing the id of the input element.}

\item{variable}{A string representing the variable to split. Default is "INS_Splits_variabele".}

\item{name}{A string representing the name. Default is "color".}

\item{label}{A string representing the label of the input. Default is "Kleur".}

\item{df}{A data frame containing the data. Default is dfCombi_geaggregeerd.}
}
\value{
A pickerInput object.
}
\description{
Function to create a picker input for splitting variables.
}
