% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_pickers_general.R
\name{pickerValues}
\alias{pickerValues}
\title{pickerValues function}
\usage{
pickerValues(
  id,
  df,
  variable = "faculty",
  role = "left",
  selected = "All",
  multiple = TRUE
)
}
\arguments{
\item{id}{A string representing the id of the input element.}

\item{df}{A data frame containing the data.}

\item{variable}{A string representing the variable to filter. Default is "faculty".}

\item{role}{A string representing the role. Default is "left".}

\item{selected}{The selected value. Default is "All".}

\item{multiple}{A boolean indicating whether multiple selections are allowed. Default is TRUE.}
}
\value{
A pickerInput object.
}
\description{
Function to create a picker input for filtering value.
}
