% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_pickers_general.R
\name{pickerVar}
\alias{pickerVar}
\title{pickerVar function}
\usage{
pickerVar(id, element, df_categories, label = NULL)
}
\arguments{
\item{id}{A string representing the id of the input element.}

\item{element}{A string representing the element.}

\item{df_categories}{A dataframe with metadata about the available categories per picker element}

\item{label}{A string representing the label of the input. Default is NULL.}
}
\value{
A pickerInput object.
}
\description{
Function to generate a picker input element based on given id and element.
}
