% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_pickers_general.R
\name{present_and_correct}
\alias{present_and_correct}
\title{present_and_correct function}
\usage{
present_and_correct(column_name, element = NA, df)
}
\arguments{
\item{column_name}{A string representing the column name.}

\item{element}{A string representing the element. Default is NA.}

\item{df}{A data frame for which to check the column. Default is dfCombi_geaggregeerd.}
}
\value{
A boolean indicating whether the column is present and correctly formed.
}
\description{
Function to check if the column is present and correctly formed based on the element type.
}
