% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_misc.R
\name{quietly_run}
\alias{quietly_run}
\title{Quietly run a function}
\usage{
quietly_run(func, ...)
}
\arguments{
\item{func}{The function to be run.}

\item{...}{Optional further arguments passed to the 'func' function.}
}
\value{
The list result of the 'func' function with messages, warnings, and output captured.
}
\description{
This function is a wrapper that allows a function to be run quietly without the need to create a separate quiet function.
}
\examples{
warning_func_arugment <- function(info) {
  warning(info)
  return("Complete")
}
result <- quietly_run(warning_func_arugment, "Just checking")
}
