% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_general.R
\name{single_module_ui}
\alias{single_module_ui}
\title{UI function for single module dashboard}
\usage{
single_module_ui(request, id, tab_item)
}
\arguments{
\item{request}{shiny request object}

\item{id}{Module id}

\item{tab_item}{Tab item UI}
}
\value{
dashboardPage UI
}
\description{
UI function for single module dashboard
}
