% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_ggplot.R
\name{stacked_composition_bar_chart}
\alias{stacked_composition_bar_chart}
\title{Stacked bar chart}
\usage{
stacked_composition_bar_chart(
  df,
  x,
  color,
  id,
  facet_name_var = rlang::sym("VIS_Groep_naam"),
  percentage = FALSE,
  wrap = FALSE
)
}
\arguments{
\item{df}{The data frame used to create the plot.}

\item{x}{The variable used on the x-axis of the plot.}

\item{color}{The variable used to color the points or bars in the plot.}

\item{id}{The identifier for selecting the data frame source.}

\item{facet_name_var}{The name of the variable used for facet wrapping.}

\item{percentage}{Logical indicating whether to create a plot in percentage mode.}

\item{wrap}{Logical indicating whether to use facet wrapping.}
}
\value{
A ggplot object.
}
\description{
Create a stacked bar chart, with optional settings for percentage (or not) and wrap (or not) modes.
}
