% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_tab_tables.R
\name{tabellenPopover}
\alias{tabellenPopover}
\title{tabellenPopover function}
\usage{
tabellenPopover(..., tabblad)
}
\arguments{
\item{...}{Arguments passed to other methods.}

\item{tabblad}{A string representing the tab.}
}
\value{
A bsPopover from the spsComps package with specified content and style.
}
\description{
Function for creating popovers on the tabs of the tables.
}
