% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_tab_tables.R
\name{tabPanelTables}
\alias{tabPanelTables}
\title{tabPanelTables function}
\usage{
tabPanelTables(id, table_one = "tabel", table_two = "tabel_twee")
}
\arguments{
\item{id}{A string representing the id.}

\item{table_one}{A string representing the first table. Default is "tabel".}

\item{table_two}{A string representing the second table. Default is "tabel_twee".}
}
\value{
A tab panel with one or two tables based on the 'id' parameter.
}
\description{
Function to define the structure of tables based on the 'id' parameter.
}
