% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_tab_tables.R
\name{tabTableTwo}
\alias{tabTableTwo}
\title{tabTableTwo function}
\usage{
tabTableTwo(id, table_one, table_two)
}
\arguments{
\item{id}{A string representing the id.}

\item{table_one}{A string representing the first table.}

\item{table_two}{A string representing the second table.}
}
\value{
A tab panel with two tables.
}
\description{
Function to create a tab panel with two tables.
}
\examples{
dummy_data1 <- data.frame(
  A = 1:5,
  B = letters[1:5]
)
dummy_dt1 <- DT::datatable(dummy_data1)
dummy_data2 <- data.frame(
X = 6:10,
Y = letters[6:10]
)
dummy_dt2 <- DT::datatable(dummy_data2)
tabTableTwo("dummy_id", dummy_dt1, dummy_dt2)
}
