% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_shinydashboardplus_adaptions.R
\name{taskItemTab}
\alias{taskItemTab}
\title{taskItemTab function}
\usage{
taskItemTab(text, tab_name = NULL, href = NULL, tabSelect = FALSE)
}
\arguments{
\item{text}{The text to display for the item.}

\item{tab_name}{The name of the tab to link to. Default is NULL.}

\item{href}{The href link for the item. If NULL, it defaults to "#".}

\item{tabSelect}{A boolean indicating whether to select the tab. Default is FALSE.}
}
\value{
An HTML list item.
}
\description{
Item for above dropdownActionMenu function.
}
\examples{
taskItemTab(text = "Selected tab", tab_name = "Tab1", tabSelect = TRUE)
taskItemTab(text = "Other tab", tab_name = "Tab2", tabSelect = FALSE)
}
