% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_filters.R
\name{transform_input}
\alias{transform_input}
\title{Transform input}
\usage{
transform_input(input)
}
\arguments{
\item{input}{A list of inputs to be transformed}
}
\value{
A list containing a column and its corresponding value for filtering
}
\description{
This function transforms a list of inputs into a column and value for filtering.
}
\examples{
input = list("A;var1", "B;var1", "C;var1")
filter_element = transform_input(input)
}
