% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_datatable.R
\name{value_callback}
\alias{value_callback}
\title{Generate Javascript for datatable cell values.}
\usage{
value_callback(data)
}
\arguments{
\item{data}{A data.frame.}
}
\value{
A character vector containing the JavaScript code.
}
\description{
This function generates Javascript code for datatable cell values. The script truncates cell values to 11 characters and adds a tooltip with the full value.
}
