% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vvshiny-package.R
\docType{package}
\name{vvshiny-package}
\alias{vvshiny}
\alias{vvshiny-package}
\title{vvshiny: Create Shiny Apps}
\description{
Helper and Wrapper functions for Shiny functionality.
}
\author{
\strong{Maintainer}: Corneel Den Hartogh \email{c.f.den.hartogh@vu.nl} (\href{https://orcid.org/0000-0001-8347-7944}{ORCID}) [copyright holder]

}
\keyword{internal}
