% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_ggplot.R
\name{wrapped_chart}
\alias{wrapped_chart}
\title{Wrapped chart}
\usage{
wrapped_chart(
  df,
  x,
  y,
  color,
  id = "bench",
  df_original,
  y_left = NULL,
  y_right = NULL,
  facet_var = rlang::sym("VIS_Groep"),
  facet_name_var = rlang::sym("VIS_Groep_naam")
)
}
\arguments{
\item{df}{The data frame used to create the plot.}

\item{x}{The variable used on the x-axis of the plot.}

\item{y}{The variable used on the y-axis of the plot.}

\item{color}{The variable used to color the points or bars in the plot.}

\item{id}{The identifier for selecting the data frame source.}

\item{df_original}{The original dataframe before summarization}

\item{y_left}{The variable used on the left y-axis when creating a comparative plot.}

\item{y_right}{The variable used on the right y-axis when creating a comparative plot.}

\item{facet_var}{The variable used for facet wrapping.}

\item{facet_name_var}{The name of the variable used for facet wrapping.}
}
\value{
A ggplot object.
}
\description{
Wrapped chart function for creating a plot based on the provided dataframe and variables.
}
