% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_misc.R
\name{display_name}
\alias{display_name}
\title{Get a user-friendly display name}
\usage{
display_name(col_name, mapping_table)
}
\arguments{
\item{col_name}{A string specifying the name of the column.}

\item{mapping_table}{An optional named list or other object that maps column names to display names. If not provided, a default is inferred from the 'id' parameter.}
}
\value{
A string containing the user-friendly name for the column.
}
\description{
This function provides a user-friendly name for a column based on a mapping table, if available.
}
