% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_filters.R
\name{filter_with_lists}
\alias{filter_with_lists}
\title{Filter with lists}
\usage{
filter_with_lists(df, filters)
}
\arguments{
\item{df}{A dataframe to be filtered}

\item{filters}{A list containing column names and their corresponding values for filtering}
}
\value{
A dataframe filtered based on the input filters
}
\description{
This function filters a dataframe using a list with column and one or more values.
}
