% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_plotly.R
\name{ggplotly_with_legend}
\alias{ggplotly_with_legend}
\title{Make ggplotly and add legend with color as title}
\usage{
ggplotly_with_legend(plot, color, id)
}
\arguments{
\item{plot}{A ggplot object.}

\item{color}{A string specifying the column name to be used as the color variable.}

\item{id}{A string specifying the ID associated with the data.}
}
\value{
A plotly object with a formatted legend.
}
\description{
This function creates a Plotly version of a ggplot2 object and adds a legend with the user-friendly name of the color variable as its title.
}
