% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_filters.R
\name{keep_values}
\alias{keep_values}
\title{Keep values}
\usage{
keep_values(input)
}
\arguments{
\item{input}{A character vector with ";"-separated strings}
}
\value{
A list of values before the semicolon in the input
}
\description{
This function extracts values before the semicolon from a ";"-separated string.
}
\examples{
input = c("A;var1", "B;var1", "C;var1")
values = keep_values(input)
}
