% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_datatable.R
\name{prep_table}
\alias{prep_table}
\title{Prepare a data table for displaying}
\usage{
prep_table(
  y,
  df,
  df_summmarized,
  id,
  y_right = NULL,
  facet_var = rlang::sym("VIS_Groep"),
  facet_name_var = rlang::sym("VIS_Groep_naam"),
  ...
)
}
\arguments{
\item{y}{A string specifying the column name to be used as the y-axis variable.}

\item{df}{A data frame containing the raw data.}

\item{df_summmarized}{A data frame containing the summarized data.}

\item{id}{A string specifying the ID associated with the data.}

\item{y_right}{An optional string specifying the column name to be used as the second y-axis variable. Default is NULL.}

\item{facet_var}{A symbol specifying the column to be used for faceting. Default is 'VIS_Groep'.}

\item{facet_name_var}{A symbol specifying the column to be used for faceting names. Default is 'VIS_Groep_naam'.}

\item{...}{Further arguments passed on to the 'make_basic_table' function.}
}
\value{
A DT::datatable object ready for displaying.
}
\description{
This function prepares a data table for displaying by providing user-friendly names, removing unneeded variables, and formatting percentages.
}
