% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_misc.R
\name{quietly_run}
\alias{quietly_run}
\title{Quietly run a function}
\usage{
quietly_run(func, ...)
}
\arguments{
\item{func}{The function to be run.}

\item{...}{Further arguments passed to the 'func' function.}
}
\value{
The result of the 'func' function with messages, warnings, and errors captured.
}
\description{
This function is a wrapper that allows a function to be run quietly without the need to create a separate quiet function.
}
