% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_tab_tables.R
\name{tabTableTwo}
\alias{tabTableTwo}
\title{tabTableTwo function}
\usage{
tabTableTwo(id, table_one, table_two)
}
\arguments{
\item{id}{A string representing the id.}

\item{table_one}{A string representing the first table.}

\item{table_two}{A string representing the second table.}
}
\value{
A tab panel with two tables.
}
\description{
Function to create a tab panel with two tables.
}
